Extract project and target environment from ARGV
os::args() returns a vector of `String`s. We know we'll always have an
0th entry, as that's how the program was executed. In the case of
environment prefix, we don't if an argument was provided.

The Vec type doesn't give us any way to get an item wrapped in an
option; we need to convert it into a slice first:

> Slices are a view into a block of memory represented as a pointer and
> a length.

Because slices return a view of memory, methods such as get() return a
borrowed version of the contents of the Vec. As far as I know this
borrow will last as long as `args` is in scope. Because the environment
can potentially last beyond `args`, we clone the String, which gives us
a copy of the String in a new area of memory. We then provide a default
string value if the environment has not been specified.